Migrate SearchTweet model to AR
