+++ app/models/carto/helpers/organization_commons.rb
@@ -1,3 +1,15 @@
+
+module Carto::OrganizationCommons
+  class OrganizationWithoutOwner < StandardError
+    attr_reader :organization
+
+    def initialize(organization)
+      @organization = organization
+      super "Organization #{organization.name} has no owner"
+    end
+  end
+end
+
 module Carto::OrganizationSoftLimits
   def soft_geocoding_limit?
     owner.try(:soft_geocoding_limit)
